Bug #29058967 CATS ALGORITHM SHOULD USE STABLE_SORT
The ties among transactions which have same weight according to C.A.T.S. algorithm should be resolved in favor of the transaction which waits longer, that is in FIFO order.
This change replaces std::sort with std::stable_sort to preserve original FIFO order of the queue in case of ties.

Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>