Bug#28678845 - SETTING FORCE MEMBERS 2ND TIME WITH VALID VALUE IS FAILING AFTER QUORUM LOSS
Problem
======================
Setting force members even with a correct value is failing after failed
attempt once when there is a quorum loss.

In a group {A,B}, B crashes. One goes to A and set in
group_replication_force_members the value "B", which is UNREACHABLE. This will
cause the group to hang, since it contains a forced invalid configuration.

Analysis
======================
Using group_replication_force_members should be a last resort situation. It
blindly installs a new configuration without validations. What happens in this
case is that A is self-evicted of the configuration. It would have to boot
itself again, but in the meantime the group is blocked with an invalid
configuration.

Fix
======================
Add an extra check in the XCom side to see if we are trying to set a node in
the list that is currently suspected to be dead or crashed.