Bug#28976533	ASSERTION `JOIN->BEST_READ < DOUBLE(1.79769313486231570814527423731704357E+308L)
introduced by WL#8652 LATERAL.
See test file for the query. We choose to merge the CTE qn and materialize
qn1.
Initially qn1 has an outer reference to top query's GRANDPARENT1, but
after merging of qn into top query, this reference becomes a
LATERAL reference (as GRANDPARENT1 is now in the same FROM clause as qn1).
When planning for the top query, we examine plan
PARENT1-GRANDPARENT1-GRANDPARENT2-PARENT2-qn1
qn1 being LATERAL now, we try to compute the cost of its materialization
(cost which is paid for every new row of GRANDPARENT1) and to compute
an average cost per access to qn1. This happens in
Optimize_table_order::lateral_derived_cost().
But table BB is empty, so the planner has prefix_rowcount==0
for all of PARENT1-GRANDPARENT1-GRANDPARENT2-PARENT2, and so the division
by 0 (to make the average) produces NaN which then causes the assertion.
Fix: avoid division by 0, report a cost of 0 for qn1.
Planning then continues. Finally query is shortcut because of
the impossible BETWEEN condition.