WL#10162 Create ndb_schema_result table
Create new hidden util table in NDB to be used for replying from
participants to the coordinator when handling a schema operation.
With the new table it's possible to return a status as well as
message indicating the result of the schema operation.

Change ndb_restore to detect the new table as a "system table". Improve
comment in BackupRestore::createSysTable() to describe what happens for
those table not listed.

Fix ndbinfo.result to adapt to the new util table

Change-Id: I4a286639433a3c9a3dec8086697719879297e075