Bug#29006668 UNEXPECTED ERROR: "ORDER BY CLAUSE IS NOT IN GROUP BY CLAUSE" WITH ROLLUP
Let's start with this working query:
 select abs(i) as a from t group by a order by a+1;
in 'a+1', 'a' gets m_alias_of_expr==true (in setup_order()) because it
is an alias to that of SELECT list. Due to this mark, the group-by-checker
just skips it, here:
Group_check::do_ident_check(
  const Bool3 local = i->local_column(select);
  if (local.is_false()) goto ignore_children;
...
ignore_children:
  stop_at(i);
  return false;
and the reason why it skips it is explained in the comment in front of
Item_ident::local_column().

Now let's add ROLLUP:
 select abs(i) as a  from t group by a with rollup order by a+1;

For a+1, between setup_order() and group-by-check, in
change_func_or_wf_group_ref() there is a replacement of 'a'
by another Item_ref and the m_alias_of_expr mark is lost.
So group-by-checker doesn't see that this 'a' is ok, and so all it
can do is looking at parts of it (i.e. 'i') and as 'i' is not in GROUP BY,
it throws an error.

Fix: propagate m_alias_of_expr when making the ROLLUP Item_ref