BUG#28601588: GROUP_REPLICATION.GR_START_STOP_APPLIER_THD_CRASH FAILS SPORADICALLY ON PB2
This test was introduced to validate that the START
GROUP_REPLICATION command could be interrupted with CTRL-C/Z'ED,
situation on which the GR applier thread is killed and depending on
the stage it was can take some time to react.
On some cases, like on slow platforms, it does take so much time
that will disallow a concurrent START GROUP_REPLICATION to take
control since the slave info tables are still under control of the
killed command. This will result in a failed concurrent START
GROUP_REPLICATION which is a correct behaviour.
Unfortunately this will make the test to fail, since it expects the
concurrent START GROUP_REPLICATION to always succeed.

To avoid the above expectation mismatch, since we already have this
kill path covered by valgrind on other tests, we will prevent this
test from running on valgrind.