Bug#28670843 -- ERROR DUPLICATE ENTRY FOR KEY PRIMARY ON DDL
Sequential applying of DDL statements CREATE TABLE/DROP TABLE against
the same table name on slave could result in the error 'duplicate entry
on primary key' on attempt to insert a record into the table mysql.tablespaces
that being done during server recovery in case a server started with the option
--slave-parallel-workers=4.

The reason of emitting the error 'duplicate entry on primary key' on slave
is that in case slave started with option --slave-parallel-workers=4 it replays
the statement CREATE TABLE without calling the method ha_innobase::open().
InnoDB depends on the call to open() to initialize the AUTOINC properly.
Without calling of the method ha_innobase::open(), the AUTOINC would be
out-dated and results in the duplicate key issue. This generic logic(calling
open() to advance AUTOINC counter for the first time, and then doing DML on
the table) must be applied to all tables residing in InnoDB, regardless of
it's a DD table or normal user table.

To fix the bug the table cache is cleaned right after the data dictionary
subsystem has been recovered via the invocation of the standalone function
close_cached_tables.

The original bug can be reproduced with the following test case:
  --let $rpl_skip_start_slave= 1
  --source include/master-slave.inc

  --source include/rpl_connection_master.inc
  CREATE TABLE t1 (c1 INT) ENGINE=InnoDB;

  --source include/rpl_connection_slave.inc
  --source include/kill_and_restart_mysqld.inc

  --let $rpl_server_number= 2

  --source include/rpl_reconnect.inc

  --source include/start_slave.inc
  --source include/rpl_connection_master.inc
  DROP TABLE t1;
  --source include/rpl_end.inc

Having this test case in the file mysql-test/t/repro.test run of
the following command
  ./mtr main.repro --mysqld=--slave-parallel-workers=4 --repeat=2
will result in the original error described at the bug report.

Since there is no way to provide the option --repeat to mysql-test-run.pl
via an option file the test case wasn't included into the patch.