Bug#29014272 THE TABLE COMMENT IS CUT DOWN ON SELECTING WITH ORDER BY
The problem is that the output of INFORMATION_SCHEMA(I_S).TABLES.TABLE_COMMENT
column shows partial comment and not full.

The root cause of the problem is that Item_str_func::max_length for
internal function implemented to project I_S.TABLES.TABLE_COMMENT, was set
to a length 256 which is less than the actual mysql.tables.comment DD
column length (2048). This patch fixes Item_str_func::max_length to 2048.

A test case is added in main.information_schema_cs.

Change-Id: Ia9c5897c9575c1dd2383c6dbd1629d266f9f1357