WL#8517 Backup parallelism : Part 3 of 25
mt-backup

Add an MT_BACKUP flag to decide the directory structure of
the backup. The MT_BACKUP flag is set to false in these
cases:
- ndbds
- ndbmtds with only one LDM worker

If the MT_BACKUP flag is set, a non-zero backup-part-ID is
passed to NDBFS so that the multithreaded backup directory
structure is used. If it is false, the old single-threaded
backup structure is used.