BUG#29361890 json_schema_embedder can not handle NUL byte
json_schema_embedded is used to compile json-schema's files into
executables.

The input file is converted into a char[] that's placed into a .cc file.

As only the .cc file is generated, the size of the char[] isn't known.

- calling code has to call strlen() to the length of the generated
  string
- json-schema-embedded MUST place a trailing 0x00 into byte-sequence

Which means the input can never contain a 0x00 byte and therefore isn't
binary-safe.

Using a static member of a class

class {name} {
public:
  static constexpr const char data_[]{ ... };
};

- ensures the size is known at compile-time
- it is binary safe

RB: 21507