WL#12616 Make number of PAUSES in spin loops configurable
Introduced new dynamic sysvar @@innodb_spin_wait_pause_multiplier
which lets you replace the default value of 50, with any number in range 0 to 100.
It controls how many times in a row to use a PAUSE instruction to achieve one unit
of delay in a spin lock (see @@innodb_spin_wait_delay), defaults to 50,
which is backward-compatible.

RB: 20851

Reviewed-by: Sunny Bains <Sunny.Bains@oracle.com>
Reviewed-by: Dimitri Kravtchuk <dimitri.kravtchuk@oracle.com>