BUG#28684376 MAKE GR REPLICATION CHANNELS ROTATE RELAY LOG ON FLUSH LOGS
Issue description
-----------------

In the context of WL#12080, relay log rotation needs to be supported, while
Group Replication is enabled, in order for binlog encryption keys to be rotated
and disposed of, properly.

Up until WL#12080, relay log rotation is disabled for Group Replication applier
channels.

Analysis
--------

After analyzing several hypothesis, the one choosen and implemented was already
described on WL#10454, which intends to prevent transaction to be split among
several relay-logs. If such property were to be observed, no changes in the
rotation and recovery processes are needed. Furthermore, it keeps intact Group
Replication behaviour - which already prevents transactions from being split.

Fix
---

For Group Replication channels, defer relay log rotation requests that are
received in the middle of a transaction being applied, to be performed after
transaction ends. The client or channel that requests the rotation should be
made wait for the actual rotation to occur.