Bug #29053813 CREATE AN XCOM UNIT TEST FOR BUG #28966455
Problem
========================================================================
As part of the analysis of Bug #28966455 APPLIER LOG MISSES A
TRANSACTION IN GR, we have identified a message flow that leads to XCom
delivering different values for the same synod.
It would be helpful to encode the diagram that represents the
problematic message flow in a XCom unit test.

Solution
========================================================================
Refactor some XCom functions that simultaneously create (process)
messages and send them (their replies) into functions that "only"
create (process) messages, and other functions that send the created
messages.

Create a unit test that uses the function that create (process) messages
to encode the diagram discussed in Bug #28966455 APPLIER LOG MISSES A
TRANSACTION IN GR.

Reviewed-by: André Negrão <andre.negrao@oracle.com>
Reviewed-by: Sujatha Sivakumar <sujatha.sivakumar@oracle.com>
RB: 21188