Bug#28954838 ASSERTION `(REMAINING_TABLES_AFTER != 0) || ((CUR_EMBEDDING_MAP == 0) && (JOIN->
It is about a query with IN subquery, and that subquery contains a
derived table 'qn' which contains an outer reference to the top query.
Semijoin transformation merges the subquery into the top query,
so 'qn' becomes LATERAL.
There is another IN subquery, with a 'qn2', which is _not_ outer-correlated,
is semi-joined and is not lateral.
In optimize_semijoin_nests_for_materialization() for the nest of 'qn2',
we have to make a plan for the sj-inner tables only. During planning,
we maintain deps_of_remaining_lateral_tables, map of dependencies of
{lateral tables not yet in plan's prefix}. In this calculation, we forgot
to exclude lateral tables out of the realm of planning, i.e. we forgot to
consider only sj-inner lateral tables ('qn2', here); we added the deps
of 'qn', even though 'qn' isn't in the nest.
So the dependencies of 'qn' were wrongly considered necessary, and an
assertion failed as the plan (which is 'qn2') didn't include them.

Fix: don't add deps of irrelevant lateral tables (found in excluded_tables).
Also removed a wrong call in optimize_straight_join(): it was wrong
to restore() as planning progresses forward and never pops a table
(unlike in best_extension_etc()).

Reviewed by: Evgeny Potemkin <evgeny.potemkin@oracle.com>