Bug#29171668: ODBC SUBSTITUTION FOR LAST_INSERT_ID() TRIGGERS A NON-DETERMINISTIC TRANSFORMATION
When the system variable sql_auto_is_null is enabled, a WHERE clause
which looks like WHERE auto_increment_col IS NULL, is rewritten to
WHERE auto_increment_col = LAST_INSERT_ID(). This transformation is
only performed once per auto-incremented value, which makes it
difficult to tell up front if the transformation will be performed.

This patch changes the transformation so that it is performed
unconditionally when sql_auto_is_null is enabled. It also fixes a bug
where the an auto-incremented value outside of the signed bigint range
would not be matched, because the original transformation treated the
value of LAST_INSERT_ID() as a signed value.

Change-Id: If97c3675267c817f45a8dcb119cc3d10a6c68bd3