WL#11999 Add support for JSON Schema
This patch adds support for caching the JSON Schema if it is const
during the execution. This gives a speedup in several normal use
cases where the JSON Schema never changes, i.e:

  SELECT
    JSON_SCHEMA_VALID('{ "type": "object" }', json_col) AS valid
  FROM
    tbl;

In the above case, the first argument will be cached so that it
can be re-used without parsing it more than once.

Change-Id: Ib944685254fda03bd774a61c06365d6020076f3e