WL#12663 ADD VERSION TO ALL TOOLS AND SCRIPTS
Adds version number to all the autotest tools namely, Atrt, ndb_cpcd, and
ndb_cpcc.

A new option '-V'/'--version' is added to atrt, ndb_cpcd, and ndb_cpcc to
display the version of MySQL, NDB and internal revision version which was not
supported earlier. In addition to this, to gain some sanity while debugging,
version number is also logged for Atrt and Cpcd whenever those processes start.
We set 1 as internal revision number for all tools. Further changes in the tool
should be accompanied by changing this value.