Bug #28965972  ATOMIC DDL + TABLE FILTER + BINLOGLESS SLAVE + GTID = DEBUG ASSERTION
Problem
=======
When applying a DDL statement on a slave, and the statement is filtered
out, and the binary log is disabled on the slave, and gtids are
enabled, the server raises an assertion in debug mode.

Analysis
========
When applying a DDL statement on a slave and the statement is filtered
out by a table filter, we report an error "ER_SLAVE_IGNORED_TABLE" to
warn slave applier thread. We need to save the DDL statement's gtid
into mysql.gtid_executed system table if the binary log is disabled
on the slave and gtids are enabled. The server asserts that there is
no error when committing the DDL statement's gtid into table. So it
raises the assertion.

Fix
===
To avoid to raise the assertion, we relax the assertion condition,
because it is not necessary to assert that there is no error in
the case.