Bug#29110613: UNICODE STRING INCORRECTLY MARKED AS INVALID
The warning message is wrong. When user runs 'SELECT x as y', 'y' is
converted to utf8 first if the charset_connection is not utf8. So if 'y' is
an valid utf8mb4 character that is not in BMP, user will get a warning
message because utf8 cannot handle it correctly. But the problem is
THD::convert_string() always uses 'from_cs' (utf8mb4) to compose the
message.

This patch doesn't fix the problem within prepared statement. It will be
fixed in separate bug thread in the future.

Fix:
Change the message to be clear. And change to raise a ERROR instead of a
warning.

Change-Id: I3d278b1b7b3b59201e6638e3f776cd94eba69d0c