Bug#29192925: DECLARE LEAF CLASSES IN THE FIELD HIERARCHY AS FINAL
Declare the classes in the Field hierarchy as final if there are no
classes inheriting from them. Similarly, declare the leaf
implementations of virtual functions in these classes as final.

By declaring them as final, it is easier for readers of the code to
see that the functions are not overridden by subclasses. It also makes
it easier for compilers to see which implementation of the function to
call in some cases, which could result in more efficient binaries with
fewer virtual function calls.

Change-Id: If19111e75751d6ae58b7dc18e6894e017e15c42a