WL#11615 - MYSQL GCS: Improve XCom Cache Management
XCom keeps the messages exchanged as a part of the consensus protocol in
a cache. Among other things, the cache is used for node recovery: if an
XCom node becomes unreachable for some time without being expelled, when
it comes back to the group it will recover the missed messages from the
caches of the other nodes.

Currently, the cache has a fixed number of entries and a fixed 1GB size.
Because of this size limit and the fact that users can define arbitrary
expel timeouts, it is possible that the caches of alive nodes become
exhausted while another node is unreachable without being expelled. If
that occurs, when the unreachable node comes back it will not be able to
recover and will be expelled, even though it respected the expel timeout.
In addition, having a fixed number of entries can make XCom too slow or
even block in high load systems, since the number of available cache
entries can quickly be depleted.

This worklog will remove the static nature of the XCom cache by allowing
users to set the size limit of the XCom cache. To that end, we will expose
a new option named group_replication_message_cache_size. In addition, the
cache will no longer be bound by a fixed number of entries; instead, it
will grow dynamically, so long as the size limit is respected.

RB (design): 20207
RB (Code): 20947