Bug#29164271 ASSERT !EXISTS FAILED AT NDB_SHARE_REFERENCES::INSERT()
Crash occured when trying to do concurrent DDL on logfile groups. The
assumption is that such DDL are serialized by holding the GSL and thus a
"hardcoded" key could be used as reference when acquiring the NDB_SHARE.
The GSL was held but not for the entire duration of the DDL operation,
thus allowing the crashing code to run concurrently.

Fix by moving the acquisition of MDL locks(and thus also the GSL lock)
on the logfile group to the caller. NOTE! Using the same function as for
locking tablespaces since the logfile group is stored as such.

Also add check that GSL is always held when entering the
ndbcluster_alter_tablespace() function.

Change-Id: Id4743e05633b8c571663b95756dbf026ffd78eeb