Bug#28723669 : RECENT REGRESSION: CRASH IN ITEM_SUBSELECT::EXEC
Problem:
For the query in bugpage, while converting subquery from IN to EXISTS
(as an optimization), a trigger condition having a subquery is added to
an always false where condition. This results in removal of the entire condition
including the subquery that was part of the trigger condition when
fix_fields is called for the where condition. However the now removed
subquery is already cached in Item_in_optimizer object created earlier.
This results in accessing a non existent select item when evaluating the
IN subquery.

Solution:
The explanation in the function
Item_in_subselect::single_value_in_to_exists_transformer
clearly states that fix fields should not modify the
condition and hence is passing NULL as reference for Item_cond::fix_fields.
So the optimization applied for ALWAYS FALSE and ALWAYS TRUE conditions
is not applied when passed reference to the function Item_cond::fix_fields
is NULL.