Bug #29214970: WRONG RESULTS FOR GROUP EXPRESSIONS WITH THE ITERATOR EXECUTOR
A late-minute fix in WL #12074 (Volcano iterator executor base) changed when
AggregateIterator calls copy_funcs(); instead of calling it on the first row
of the group, it calls it when we've seen the end of the group (ie., when the
next row has been seen). This fixes the case of items depending on
aggregates, such as SUM(f1) + 1, but it breaks expressions on group
expressions, such as SELECT f1 + 1 GROUP BY f1, when materialization is due
to ORDER BY.

Usually, this was masked, as AggregateIterator is rarely used in such
queries; TemptableAggregateIterator is preferred instead (probably wrongly in
many cases, but that's a different issue). However, it is possible to force
the optimizer into creating such a plan by using an aggregate with DISTINCT
(or ROLLUP, which was what originally exposed the bug).

To fix it, we call copy_funcs() at the beginning as we did before said fix.
To get the correct results for the items depending on aggregates, we copy them
separately at the end; they have already been split to depend on refs in
setup_fields() because ORDER BY needs it, so any other inputs should be fine.

Cherry-pick from trunk to 8.0.

Change-Id: Ia4c1526ba7ee1807848cc0909b20dc167caaafd7