WL#10137: REDO log encryption support in MEB-8.0
When the redo log of an InnoDB server instance was encrypted, the
backup and apply-log operations failed because MEB did not decrypt
redo log. During backup and apply-log MEB looks inside individual redo
log records, and if the Server encrypts redo log, these log records
are encrypted. In this case, MEB must decrypt all redo log records
that it inspects.

The encryption of redo log is based on the same two-level key system
as for TDE tables. The encryption information for redo log is stored
in the header of "ib_logfile0". The encryption information stored in
this file consists of master key id, an initialization vector and a
tablespace key. The latter two are encrypted with the master key.

This patch modifies MEB to decrypt (temporarily) redo log during
backup and apply-log when that is required. This patch also modifies
the backup by including new optional "redo_log_encryption_info"
attribute in the "backup_variables.txt" file.

Summary of the major changes:

1) Modify Redo_log_processor::run() to decrypt encrypted redo log
   before it is processed during backup and verify the block log checksums.

2) Add decryption of redo log in Back_apply_log::process()
   for apply-log.

3) Add a new optional "redo_log_encryption_info" variable in the
   "backup_variables.txt". This value of this variable is the
   encryption information copied from the header of "ib_logfile0"
   during backup. This saved encryption information is used in
   apply-log.

4) Initialization routines for setting up redo log decryption:
   Redo_log_adapter::init_encryption_for_backup() and
   Redo_log_adapter::init_encryption_for_apply_log().

5) Add loading of master keys for backup:
   TDE_keyring_service::load_keys_for_backup().

6) Add MEB code in InnoDB source for reading encryption information
   from "ib_logfile0" and decoding redo log:
   fil_redo_io() and meb_read_log_encryption.

7) Add a method for checking if a sequence redo log blocks contains
   any encrypted log blocks:
   Redo_log_adapter::contains_encrypted_blocks().

8) Modify Redo_log_adapter::scan_buffer() and meb_scan_log_seg() to
   allow excluding optionally checksum verification.

9) Add CPU time measurement for redo log decryption during backup
   for performance testing (behind #ifdef DEBUG_REDO_LOG_DECRYPTION).

This patch fixes also Bug#27771552.

RB: 20934