Bug#29215410: USE INDEXES TO ESTIMATE FILTER EFFECT IN SECONDARY ENGINES
Enable use of indexes in the optimizer when optimizing statements for
execution in a secondary storage engine. Secondary storage engines do
not have indexes, so the optimizer will instead get the relevant
information from the indexes in the primary storage engine.

The use of indexes enabled by this patch is limited to retrieving
estimates for the size of a range when an indexed column is compared
to a known value. That is, in queries such as

  SELECT * FROM t WHERE indexed_col = 123;

  SELECT * FROM t WHERE indexed_col < 123;

  SELECT * FROM t WHERE indexed_col BETWEEN 123 AND 456;

the optimizer will now use the estimate provided by the
records_in_range() function of the secondary storage engine handler
to determine how many rows will match the predicate.

To take advantage of indexes during optimization, the secondary engine
storage handler needs to do the following:

1) Provide an implementation of handler::records_in_range() that gets
the estimate from the primary storage engine. The primary storage
engine (InnoDB) will typically estimate the size of a range by using
index dives.

2) Signal via its handler::index_flags() function that it can estimate
range sizes.

3) Signal via its handler::table_flags() function that it cannot use
indexes for access.

Change-Id: I8e00bb88c92837699f5b49bcba898628c7ac8b46