Bug#29155439 ASSERTION FALSE IN WIDEN_FRACTION() DECIMAL.CC, LINE 993
The logic for truncating/extending a decimal constant to the desired
number of fractional digits during constant folding was lacking. Extra
trailing zeros in the fraction could trigger an attempt to widen the
fraction, since the decision to widen was based on the number of
non-zero fractional digits, rather than on total number of fractional
digits. The widen_fraction method would then hit the assert seen in
the repro.

Solution: identify where we can just truncate excess trailing
zeros. In this case we do not need to adjust the comparison operator,
we just replace the constant with one with a smaller number of
trailing zeros in the fraction. This also avoids us trying to call
widen_fraction with wrong arguments.

Change-Id: Id4e3c3042e2479d64c1968713b5a54dc49120685