Bug#28949332 SIG6-TEMPTABLE::HANDLER::PRIMARY_KEY_IS_CLUSTERED () AT SRC/HANDLER.CC
Problem:
The method temptable::handler::primary_key_is_clustered implements the
virtual method in the parent class. However, this method is not
implemented in the class temptable and as per the current
implementation, calling this function hits an assert in the debug build.

Solution:
Instead of hitting the assert, it is ok to just return false.

Fix:
Removed the assert from the code.

RB: 21232
Reviewed by: Mayank Prasad (mayank.prasad@oracle.com)