Bug#29231490: REGEXP FUNCTIONS FAIL TO USE CODEPOINT POSITIONS
Adds functions that convert positions as seen by the user,
expected to refer to codepoints, to ICU's internally used
UTF-16 words, and convert such positions as returned by
ICU to codepoint number.

We also optimize so that we don't convert prefixes of the
subject string that won't be searched by the regexp engine.

Change-Id: I8287145cb0efc436c5b6e0f148f61a57c7b84849