Bug#29157265: ASSERTION FAILURE:    SRV_SHUTDOWN_STATE <= SRV_SHUTDOWN_FLUSH_PHASE
This bug occurred because of a circular relative path where the path
contains "/anydir/../".  The directory used was a database directory.
This script cause an immediate shutdown.
  CREATE DATABASE anydir;
  CREATE TABLESPACE ts ADD DATAFILE 'anydir/../ts.ibd';
  DROP DATABASE anydir;

The file handle is opened for writing by the background IO thread after
the CREATE TABLESPACE completes. So if the DROP DATABASE occurs first
this path cannot be opened on Linux. It works on Windows. Since the IO
thread needs to write dirty pages, it must cause an immediate shutdown.
So os_file_handle_error_cond_exit() calls srv_fatal_error() which uses
srv_shutdown_all_bg_threads().  While this function is waiting for other
background threads to exit, the purge thread calls log_buffer_reserve()
which hits this assert:
     ut_a(srv_shutdown_state <= SRV_SHUTDOWN_FLUSH_PHASE);
when it changes an undo page and needs to commit an MTR.

This immediate shutdown is caused by an unexpected invalid path.
The shutdown can be caused by three simple SQL statements.
The assert that happens is an inability to shutdown cleanly.
But the immediate shutdown occurs either way.

The solution is to reject circular paths segments.

There is also code in this patch to allow a cleaner shutdown when an
immediate shutdown like this must occur.

Approved by Pawel in RB#21340