BUG#28799939 DISALLOW ALTER TABLE ADD COLUMN WITH NONDETERMINISTIC DEFAULT, IF BINLOG ENABLED
Description
-----------

The execution of `ALTER TABLE ... ADD COLUMN ... DEFAULT (NDF function)` should
be allowed when the binlog is enabled and `BINLOG_FORMAT=STATEMENT`, independently
of `GTID_MODE` or `ENFORCE_GTID_CONSISTENCY` values.

Analysis
--------

The validation for the DDL statement safety is not correctly placed in
`THD::is_ddl_gtid_compatible()` since GTID consistency is not the correct
dependency but rather if it is possible to safely replicate the statement with a
NDF as the `DEFAULT` expression.

Fix
---

1) In order to keep the implementation logic consistent, move the unsafe
   statement validation from the GTID consistency validation into the `ALTER
   TABLE` statement execution flow - `mysql_alter_table()` function.

2) Always trigger a warning when binlog is enabled and the statement is allowed
   to be executed - `BINLOG_FORMAT=STATEMENT` - both to the server error log and
   to the client channel.

ReviewBoard: 21234
Reviewed-by: Daogang Qu <bill.qu@oracle.com>
Reviewed-by: Neha Kumari <neha.n.kumari@oracle.com>