Bug#29240446: RECENT REGRESSION, ISCTYPE.CPP(36) : ASSERTION FAILED: C >= -1 && C <= 255
The problem was that functions that converted strings to date types
could cause an assertion on Windows debug builds. The root cause
was that the behavior of <cctype> functions such as isdigit() is
undefined if their input is neither representable as unsigned char
nor equal to EOF. The patch for Bug#29041740 introduced usage of
these functions where the argument was plain 'char'.

This patch fixes the problem by casting <cctype> function input
to unsigned char.

Change-Id: I7bfcb1e22b794e2ae1d1354c12402c5563462c76