Bug#29251523: CHANGE THE SIGNATURE OF FIELD::GET_PTR()
Make Field::get_ptr() return the pointer instead of taking an output
argument that it sets to the pointer. This makes the calling code
simpler, as it allows to do declaration and assignment in a single
statement, and sometimes allows elimination of intermediate variables
altogether. Also make the returned pointer const.

Change-Id: Iefaa21684cdc8a67631904940a4968ff8b93e567