WL#12704 REMOVE LIBNDBCLIENT DEPENDENCY FROM NDB_CPCC
Removes the dependency of ndb_cpcc on libndbclient shared library.

ndb_cpcc is one of the autotest tools which is supposed to run
independently on autotest server. But, the latest version of ndb_cpcc
was dependent on the libndbclient shared library which introduced the
constraint of having a set $LD_LIBRARY_PATH. Now, ndb_cpcc is made to
make use of ndbclient_static library thereby removing the dependency on
the shared library.

Approved by: Tiago Alves <tiago.alves@oracle.com>