Bug#29273194: FAILURE RUNNING TESTS INNODB.XA_RECOVERY_DEBUG before INNODB_UNDO.UNDO_DDL_VS_DML
The test innodb.xa_recovery_debug is introduced a fix for
   Bug#20872655 XA ROLLBACK IS NOT CRASH-SAFE
that introduced a bug not notices until the new test
innodb_undo.undo_ddl_vs_dml is run after it.

The innodb.xa_recovery_debug test crashes while an XA transaction
is in the process of being rolled back.  The followig restart and
recovery finds two undo log entries to rollback.  That is, two
changes in the same XA transaction.  The first one happened when
it was prepared and the second when it is put back into active
state.  This second undo log entry was the fix that made XA rollback
crash safe.

During recovery, the routine called trx_resurrect() is provided with
an rseg to attach to the transactions that need to be resureected.
It does this in both trx_resurrect_insert() and trx_resurrect_update.

When a transaction contains both an insert and an update, both of
these attach that same rseg to the transaction.  The second assignment
overwrites the first so that is not bad.  But both routines also
increment rseg->trx_ref_count.

Now rseg->trx_ref_count is one higher than the number of transactions
that use that rseg.

This causes a problem for undo tablespace truncation.  The routine
that determines if the undo space is empty finds an rseg with nothing
in it while the rseg->trx_ref_count is still set to 1.
So the undo space is stuck in inactive mode until the next reboot.

Whats more, when this happens, the purge thread is extra busy checking
for this space to become empty. and it never will until a reboot occurs.

The new test, innodb_undo.undo_ddl_vs_dml, does a lot of DML which
causes innodb_undo_001 grow enouph that it needs to be truncated.
Then this undo space gets stuck in inactive mode.  This happens in
innodb_undo.undo_ddl_vs_dml when innodb.xa_rollback_debug is run
before it. Then when undo_ddl_vs_dml tries to truncate undo_003,
it never succeeds because the purge thread is still trying to truncate
innodb_undo_001.

The fix for this bug is to change trx_resurrect_update() so that it
notices when a transaction already has an rseg attached.  It will
assert that the attached rseg is the same one it was about to attach,
then skip the rseg->trx_ref_count++ so that the rseg->trx_ref_count
is correct.

Approved by Bin in RB#21421