Bug#29240665 ASAN REPORTS MEMORY LEAKS ON SECONDARY ENGINE MASTER DURING SHUTDOWN AFTER LOAD AND UNLOAD
Issue
=====
a) Each context that a thread works with has a range that it needs to process
and each range maintians a start row and end row information. In case of
cleanup, for a non-partitioned table the logic used is that the start row of
the first range would be freed initially and then the end row of the subsequent
ranges would be freed which ensures that all the rows are freed as the start
row of all the ranges would be the same as the end rows of all the previous
ranges. The first part of the logic does not work in the case of partitioned
table, in secondary engine context, as there could be many ranges (the first
range in all the partitions) where the first row needs to be freed.

b) Also, there was this case where the splitting was happening twice - once in
ha_innopart::pread_adapter_parallel_scan_start() when we try to find out the
number of threads that needs to be spawned and then again while we're doing the
actual read.

Fix
===
a) I've fixed the first part of the logic by explicitly calling destroy on the
start row of the first range in all the partitions.

b) In calc_num_threads() we check if the partitions have already been created
and if it is the case we just return

Reviewed-by: Sunny Bains <Sunny.Bains@oracle.com>
Change-Id: I7737a6e7792d63ec1b43919832825c18ba8ba27e